+++ commons-cache/src/main/java/com/hb0730/commons/cache/support/serial/impl/Jackson2JsonCacheWrapperSerializer.java
@@ -1,6 +1,8 @@
 package com.hb0730.commons.cache.support.serial.impl;
 
+import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.hb0730.commons.cache.CacheWrapper;
 import com.hb0730.commons.cache.exception.SerializationException;
 import com.hb0730.commons.cache.support.serial.AbstractSerializer;
@@ -16,16 +18,20 @@ import java.io.ByteArrayOutputStream;
  * @author bing_huang
  * @since 1.0.0
  */
-public class Jackson2JsonCacheWrapperSerializer extends AbstractSerializer {
-    public static final Jackson2JsonCacheWrapperSerializer JSON_STRING_SERIALIZER = new Jackson2JsonCacheWrapperSerializer(true);
+public class Jackson2JsonCacheWrapperSerializer<T> extends AbstractSerializer<T> {
     @Getter
     private final ObjectMapper objectMapper;
 
     public static final int IDENTITY_NUMBER = 0x4A953A81;
 
 
-    public Jackson2JsonCacheWrapperSerializer(boolean useIdentityNumber) {
-        super(useIdentityNumber, IDENTITY_NUMBER);
+    public Jackson2JsonCacheWrapperSerializer(boolean useIdentityNumber, Class<T> type) {
+        super(useIdentityNumber, IDENTITY_NUMBER, type);
+        this.objectMapper = new ObjectMapper();
+    }
+
+    public Jackson2JsonCacheWrapperSerializer(boolean useIdentityNumber, JavaType type) {
+        super(useIdentityNumber, IDENTITY_NUMBER, type);
         this.objectMapper = new ObjectMapper();
     }
 
@@ -36,21 +42,22 @@ public class Jackson2JsonCacheWrapperSerializer extends AbstractSerializer {
      * @param useIdentityNumber 是否使用使用标识号{@link #IDENTITY_NUMBER}
      * @since 2.0.2
      */
-    public Jackson2JsonCacheWrapperSerializer(boolean useIdentityNumber, ObjectMapper objectMapper) {
-        super(useIdentityNumber, IDENTITY_NUMBER);
+    public Jackson2JsonCacheWrapperSerializer(boolean useIdentityNumber, ObjectMapper objectMapper, Class<T> type) {
+        super(useIdentityNumber, IDENTITY_NUMBER, type);
         this.objectMapper = objectMapper;
     }
 
     @Override
-    protected Object doDeserialize(@Nullable byte[] buffer) throws Exception {
+    protected CacheWrapper<T> doDeserialize(@Nullable byte[] buffer) throws Exception {
         if (null == buffer || 0 == buffer.length) {
             return null;
         }
         try {
+            JavaType javaType = TypeFactory.defaultInstance().constructParametricType(CacheWrapper.class, getJavaType());
             if (useIdentityNumber) {
-                return getObjectMapper().readValue(buffer, 4, buffer.length - 4, CacheWrapper.class);
+                return getObjectMapper().readValue(buffer, 4, buffer.length - 4, javaType);
             }
-            return getObjectMapper().readValue(buffer, CacheWrapper.class);
+            return getObjectMapper().readValue(buffer, javaType);
 
         } catch (Exception e) {
             throw new SerializationException("jackson deserialize error:" + e.getMessage(), e);