+++ commons-cache/src/test/java/com/hb0730/commons/cache/config/CacheConfigTest.java
@@ -1,0 +1,13 @@
+package com.hb0730.commons.cache.config;
+
+import com.hb0730.commons.cache.support.serial.Serializer;
+import org.junit.Test;
+
+public class CacheConfigTest {
+
+    @Test
+    public void serializerTest() {
+        CacheConfig<String, String> config = new CacheConfig<>();
+        Serializer<String> serializer = config.getSerializer();
+    }
+}